# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_confirm_upload_api_v1_upload_confirm_post": {
        "properties": {
          "column_mapping": {
            "description": "JSON object: canonical_field \u2192 original_column_name",
            "title": "Column Mapping",
            "type": "string"
          },
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file",
          "column_mapping"
        ],
        "title": "Body_confirm_upload_api_v1_upload_confirm_post",
        "type": "object"
      },
...
      "Body_parse_upload_api_v1_upload_parse_post": {
        "properties": {
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_parse_upload_api_v1_upload_parse_post",
        "type": "object"
      },
...
      "ConfirmResponse": {
        "description": "Result of the /confirm step returned after saving the normalised portfolio.",
        "properties": {
          "enriched_count": {
            "title": "Enriched Count",
            "type": "integer"
          },
          "enrichment_details": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Enrichment Details",
            "type": "array"
          },
          "enrichment_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enrichment Note"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "holdings_parsed": {
            "title": "Holdings Parsed",
            "type": "integer"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "rows_accepted": {
            "title": "Rows Accepted",
            "type": "integer"
          },
          "rows_fully_enriched": {
            "title": "Rows Fully Enriched",
            "type": "integer"
          },
          "rows_no_fundamentals": {
            "title": "Rows No Fundamentals",
            "type": "integer"
          },
          "rows_partially_enriched": {
            "title": "Rows Partially Enriched",
            "type": "integer"
          },
          "rows_rejected": {
            "title": "Rows Rejected",
            "type": "integer"
          },
          "rows_sector_unknown": {
            "title": "Rows Sector Unknown",
            "type": "integer"
          },
          "skipped_details": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Skipped Details",
            "type": "array"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "filename",
          "rows_accepted",
          "rows_rejected",
          "skipped_details",
          "rows_fully_enriched",
          "rows_partially_enriched",
          "rows_sector_unknown",
          "rows_no_fundamentals",
          "enriched_count",
          "enrichment_note",
          "enrichment_details",
          "holdings_parsed",
          "message"
        ],
        "title": "ConfirmResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PortfolioDeltaResponse": {
        "properties": {
          "added_tickers": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Added Tickers",
            "type": "array"
          },
          "captured_at_a": {
            "format": "date-time",
            "title": "Captured At A",
            "type": "string"
          },
          "captured_at_b": {
            "format": "date-time",
            "title": "Captured At B",
            "type": "string"
          },
          "days_apart": {
            "title": "Days Apart",
            "type": "integer"
          },
          "decreased_tickers": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Decreased Tickers",
            "type": "array"
          },
          "has_changes": {
            "default": false,
            "title": "Has Changes",
            "type": "boolean"
          },
          "holding_deltas": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/HoldingDelta"
            },
            "title": "Holding Deltas",
            "type": "array"
          },
          "increased_tickers": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Increased Tickers",
            "type": "array"
          },
          "removed_tickers": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Removed Tickers",
            "type": "array"
          },
          "sector_deltas": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SectorDelta"
            },
            "title": "Sector Deltas",
            "type": "array"
          },
          "snapshot_a_id": {
            "title": "Snapshot A Id",
            "type": "integer"
          },
          "snapshot_b_id": {
            "title": "Snapshot B Id",
            "type": "integer"
          },
          "total_pnl_delta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Pnl Delta"
          },
          "total_value_delta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Value Delta"
          },
          "total_value_delta_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Value Delta Pct"
          },
          "unchanged_tickers": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Unchanged Tickers",
            "type": "array"
          }
        },
        "required": [
          "snapshot_a_id",
          "snapshot_b_id",
          "captured_at_a",
          "captured_at_b",
          "days_apart"
        ],
        "title": "PortfolioDeltaResponse",
        "type": "object"
      },
...
      "SnapshotCreateRequest": {
        "properties": {
          "label": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Human-readable label for this snapshot, e.g. 'Before rebalance'",
            "title": "Label"
          }
        },
        "title": "SnapshotCreateRequest",
        "type": "object"
      },
      "SnapshotDetail": {
        "description": "Full snapshot including per-holding rows and JSON blobs.",
        "properties": {
          "captured_at": {
            "format": "date-time",
            "title": "Captured At",
            "type": "string"
          },
          "holdings": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SnapshotHoldingRow"
            },
            "title": "Holdings",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "num_holdings": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Num Holdings"
          },
          "portfolio_id": {
            "title": "Portfolio Id",
            "type": "integer"
          },
          "risk_metrics": {
            "additionalProperties": {
              "type": "number"
            },
            "default": {},
            "title": "Risk Metrics",
            "type": "object"
          },
          "sector_weights": {
            "additionalProperties": {
              "type": "number"
            },
            "default": {},
            "title": "Sector Weights",
            "type": "object"
          },
          "top_holdings": {
            "default": [],
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Top Holdings",
            "type": "array"
          },
          "top_sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top Sector"
          },
          "total_cost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Cost"
          },
          "total_pnl": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Pnl"
          },
          "total_pnl_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Pnl Pct"
          },
          "total_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Value"
          }
        },
        "required": [
          "id",
          "portfolio_id",
          "captured_at"
        ],
        "title": "SnapshotDetail",
        "type": "object"
      },
...
      "SnapshotSummary": {
        "description": "Lightweight \u2014 for listing snapshots without loading all holdings.",
        "properties": {
          "captured_at": {
            "format": "date-time",
            "title": "Captured At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "num_holdings": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Num Holdings"
          },
          "portfolio_id": {
            "title": "Portfolio Id",
            "type": "integer"
          },
          "top_sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top Sector"
          },
          "total_cost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Cost"
          },
          "total_pnl": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Pnl"
          },
          "total_pnl_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Pnl Pct"
          },
          "total_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Value"
          }
        },
        "required": [
          "id",
          "portfolio_id",
          "captured_at"
        ],
        "title": "SnapshotSummary",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "P-Insight Portfolio Analytics API. Visit /docs for interactive API documentation.",
    "title": "P-Insight API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/portfolios/{portfolio_id}/snapshot": {
      "post": {
        "description": "Capture the current state of a portfolio as a snapshot.",
        "operationId": "create_snapshot_api_v1_portfolios__portfolio_id__snapshot_post",
        "parameters": [
          {
            "in": "path",
            "name": "portfolio_id",
            "required": true,
            "schema": {
              "title": "Portfolio Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/SnapshotCreateRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create a portfolio snapshot",
        "tags": [
          "Snapshots"
        ]
      }
    },
    "/api/v1/portfolios/{portfolio_id}/snapshots": {
      "get": {
        "operationId": "list_snapshots_api_v1_portfolios__portfolio_id__snapshots_get",
        "parameters": [
          {
            "in": "path",
            "name": "portfolio_id",
            "required": true,
            "schema": {
              "title": "Portfolio Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SnapshotSummary"
                  },
                  "title": "Response List Snapshots Api V1 Portfolios  Portfolio Id  Snapshots Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List snapshots for a portfolio",
        "tags": [
          "Snapshots"
        ]
      }
    },
...
    "/api/v1/snapshots/{snapshot_a_id}/delta/{snapshot_b_id}": {
      "get": {
        "description": "Returns a structured comparison of snapshot_a (older) vs snapshot_b (newer).\nUse snapshot_a_id=latest-1 and snapshot_b_id=latest for \"what changed\" analysis.",
        "operationId": "snapshot_delta_api_v1_snapshots__snapshot_a_id__delta__snapshot_b_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_a_id",
            "required": true,
            "schema": {
              "title": "Snapshot A Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "snapshot_b_id",
            "required": true,
            "schema": {
              "title": "Snapshot B Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioDeltaResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compute delta between two snapshots",
        "tags": [
          "Snapshots"
        ]
      }
    },
    "/api/v1/snapshots/{snapshot_id}": {
      "delete": {
        "operationId": "delete_snapshot_api_v1_snapshots__snapshot_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "title": "Snapshot Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete a snapshot",
        "tags": [
          "Snapshots"
        ]
      },
      "get": {
        "operationId": "get_snapshot_api_v1_snapshots__snapshot_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "title": "Snapshot Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get full snapshot detail",
        "tags": [
          "Snapshots"
        ]
      }
    },
...
    "/api/v1/upload/confirm": {
      "post": {
        "description": "Step 2 of the upload wizard.\n\nAccepts the file again plus the confirmed column mapping (possibly edited by the user).\nNormalises all rows, validates them, persists them to the database,\nand saves a canonical CSV for compatibility.\n\nThe column_mapping JSON looks like:\n  {\"ticker\": \"Symbol\", \"name\": \"Company Name\", \"quantity\": \"Qty\", ...}",
        "operationId": "confirm_upload_api_v1_upload_confirm_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_confirm_upload_api_v1_upload_confirm_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import portfolio from upload",
        "tags": [
          "Upload"
        ]
      }
    },
    "/api/v1/upload/parse": {
      "post": {
        "description": "Step 1 of the upload wizard.\n\nReads the file, auto-detects column roles, returns a preview.\nNo data is saved at this stage.",
        "operationId": "parse_upload_api_v1_upload_parse_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_parse_upload_api_v1_upload_parse_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Parse and detect columns",
        "tags": [
          "Upload"
        ]
      }
    },
    "/api/v1/upload/status": {
      "get": {
        "description": "Spec-required query-param variant of the enrichment status endpoint.\n\nReturns per-holding enrichment state from the DB.\nIdentical response shape to GET /upload/v2/status/{portfolio_id}.\nPoll after POST /upload/v2/confirm to track background enrichment progress.",
        "operationId": "get_upload_status_api_v1_upload_status_get",
        "parameters": [
          {
            "description": "ID of the portfolio to check",
            "in": "query",
            "name": "portfolio_id",
            "required": true,
            "schema": {
              "description": "ID of the portfolio to check",
              "title": "Portfolio Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2StatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Poll enrichment status for an uploaded portfolio (spec-compliant URL)",
        "tags": [
          "Upload"
        ]
      }
    },
...
  }
}